+++ src/app/pages/OutfittingPage.jsx
@@ -17,6 +17,7 @@ import UtilitySlotSection from '../components/UtilitySlotSection';
 import OffenceSummary from '../components/OffenceSummary';
 import DefenceSummary from '../components/DefenceSummary';
 import MovementSummary from '../components/MovementSummary';
+import DamageDealt from '../components/DamageDealt';
 import LineChart from '../components/LineChart';
 import PowerManagement from '../components/PowerManagement';
 import CostSection from '../components/CostSection';
@@ -348,30 +349,8 @@ export default class OutfittingPage extends Page {
           <MovementSummary ship={ship} code={code}/>
         </div>
 
-        <div className='group half'>
-          <table style={{ width: '100%', lineHeight: '1em', backgroundColor: 'transparent' }}>
-            <tbody >
-              <tr>
-                <td style={{ verticalAlign: 'top', padding: 0, width: '2.5em' }} onMouseEnter={termtip.bind(null, 'fuel level')} onMouseLeave={hide}>
-                  <Fuel className='xl primary-disabled' />
-                </td>
-                <td>
-                  <Slider
-                    axis={true}
-                    onChange={this._fuelChange}
-                    axisUnit={translate('T')}
-                    percent={fuelLevel}
-                    max={fuelCapacity}
-                    scale={sizeRatio}
-                    onResize={onWindowResize}
-                  />
-                </td>
-                <td className='primary' style={{ width: '10em', verticalAlign: 'top', fontSize: '0.9em', textAlign: 'left' }}>
-                  {formats.f2(fuelLevel * fuelCapacity)}{units.T} {formats.pct1(fuelLevel)}
-                </td>
-              </tr>
-            </tbody>
-          </table>
+        <div>
+          <DamageDealt ship={ship} code={code} currentMenu={menu}/>
         </div>
 
       </div>
@@ -406,4 +385,28 @@ export default class OutfittingPage extends Page {
 //            func={state.speedChartFunc}
 //          />
 //        </div>
-
+//        <div className='group half'>
+//          <table style={{ width: '100%', lineHeight: '1em', backgroundColor: 'transparent' }}>
+//            <tbody >
+//              <tr>
+//                <td style={{ verticalAlign: 'top', padding: 0, width: '2.5em' }} onMouseEnter={termtip.bind(null, 'fuel level')} onMouseLeave={hide}>
+//                  <Fuel className='xl primary-disabled' />
+//                </td>
+//                <td>
+//                  <Slider
+//                    axis={true}
+//                    onChange={this._fuelChange}
+//                    axisUnit={translate('T')}
+//                    percent={fuelLevel}
+//                    max={fuelCapacity}
+//                    scale={sizeRatio}
+//                    onResize={onWindowResize}
+//                  />
+//                </td>
+//                <td className='primary' style={{ width: '10em', verticalAlign: 'top', fontSize: '0.9em', textAlign: 'left' }}>
+//                  {formats.f2(fuelLevel * fuelCapacity)}{units.T} {formats.pct1(fuelLevel)}
+//                </td>
+//              </tr>
+//            </tbody>
+//          </table>
+//        </div>