+++ ChangeLog.md
@@ -13,6 +13,7 @@
   * Obey restricted slot rules when adding all for internal slots
   * Version URLs to handle changes to ship specifications over time
   * Do not include disabled shield boosters in calculations
+  * Add 'Damage dealt' section
 
 #2.2.5
   * Calculate rate of fire for multi-burst weapons